refactor: InputHandler 인터페이스 분리
- 기존에는 바로 클래스를 사용하여 테스트대역을 만들기 어려웠습니다.
- 따로 인터페이스를 두어서 테스트대역이 이를 구현할 수 있도록 리팩터링하였습니다.
