+++ src/main/java/lotto/io/inputHandler/InputHandler.java
@@ -1,0 +1,12 @@
+package lotto.io.inputHandler;
+
+import java.util.List;
+
+public interface InputHandler {
+
+    int handlePurchaseCost();
+
+    List<Integer> handleWinningNumbers();
+
+    int handleBonusNumber();
+}
